# perf: optimize state machine with Top state fast path